mnist_eager to use GPU if not specified otherwise (#4216)
Update mnist_eager.py to use GPU if available and not specified otherwise. At the moment it uses the CPU even if a GPU is available, since "if flags_obj.no_gpu or tf.test.is_gpu_available():" in "def run_mnist_eager(flags_obj):" evaluates to True if a GPU is available.